renamed icons + use base icons in other components
