--- packages/uui-base/lib/svgs/icon-check.ts
@@ -1,5 +1,0 @@
-import { svg } from 'lit';
-
-export default svg`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
-<path d="M461.884 68.14c-132.601 81.297-228.817 183.87-272.048 235.345l-105.874-82.95-46.751 37.691 182.941 186.049c31.485-80.646 131.198-238.264 252.956-350.252L461.884 68.14z"/>
-</svg>`;
