--- packages/uui-base/lib/svgs/icon-wrong.ts
@@ -1,5 +1,0 @@
-import { svg } from 'lit';
-
-export default svg`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
-<path d="M422.952 371.305L307.064 255.418l115.884-115.887-51.722-51.723L255.34 203.693 139.457 87.812l-51.726 51.719 115.885 115.885L87.731 371.305l51.726 51.721L255.344 307.14l115.884 115.882z"/>
-</svg>`;
