+++ packages/uui-base/lib/svgs/iconCheck.ts
