+++ packages/uui-base/lib/svgs/iconWrong.ts
@@ -1,9 +1,5 @@
 import { svg } from 'lit';
 
-export const iconCheck = svg`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
-<path d="M461.884 68.14c-132.601 81.297-228.817 183.87-272.048 235.345l-105.874-82.95-46.751 37.691 182.941 186.049c31.485-80.646 131.198-238.264 252.956-350.252L461.884 68.14z"/>
-</svg>`;
-
 export const iconWrong = svg`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
 <path d="M422.952 371.305L307.064 255.418l115.884-115.887-51.722-51.723L255.34 203.693 139.457 87.812l-51.726 51.719 115.885 115.885L87.731 371.305l51.726 51.721L255.344 307.14l115.884 115.882z"/>
 </svg>`;
