+++ packages/uui-base/lib/svgs/index.ts
@@ -1,0 +1,2 @@
+export * from './iconCheck';
+export * from './iconWrong';
