+++ packages/uui-toggle/lib/uui-toggle.element.ts
@@ -3,7 +3,7 @@ import {
   UUIHorizontalShakeKeyframes,
   UUIHorizontalShakeAnimationValue,
 } from '@umbraco-ui/uui-base/lib/animations';
-import { iconWrong, iconCheck } from './toggle-icons';
+import { iconWrong, iconCheck } from '@umbraco-ui/uui-base/lib/svgs';
 import { UUIBooleanInputBaseElement } from '@umbraco-ui/uui-base/lib/uui-boolean-input';
 
 /**
